refactor: cleanup and organize
