+++ src/spoon/get-decoder-state-callbacks.ts
