+++ src/spoon/internals-extract.ts
